setup and add test to / endpoint
