+++ package.json
@@ -2,9 +2,10 @@
   "name": "dadosjusbr-parser",
   "version": "1.0.0",
   "description": "A http API that parses judges payments .xls or .xlsx spreadsheets into a nice csv",
-  "main": "app.js",
+  "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node index.js",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -13,7 +14,7 @@
   "keywords": [
     "dadosjusbr"
   ],
-  "author": "Vinicius A Agostini",
+  "author": "dadosjusbr.online",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/dadosjusbr/parser/issues"
@@ -21,5 +22,9 @@
   "homepage": "https://github.com/dadosjusbr/parser#readme",
   "dependencies": {
     "express": "^4.16.4"
+  },
+  "devDependencies": {
+    "jest": "^23.6.0",
+    "supertest": "^3.3.0"
   }
 }
