+++ src/server.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const app = express();
+
+const parserService = require('./parser.service');
+
+app.get('/', parserService.getParsedSpreadsheet);
+
+module.exports = app;
